+++ lib/rpush.rb
@@ -1,14 +1,19 @@
+require 'logger'
 require 'multi_json'
 require 'active_support/all'
 
 module Rpush
   def self.attr_accessible_available?
     require 'rails'
+  rescue LoadError
+    false
+  else
     ::Rails::VERSION::STRING < '4' || defined?(::ActiveRecord::MassAssignmentSecurity)
   end
 end
 
 require 'rpush/version'
+require 'rpush/cli'
 require 'rpush/deprecation'
 require 'rpush/deprecatable'
 require 'rpush/logger'
@@ -24,12 +29,15 @@ module Rpush
     defined? JRUBY_VERSION
   end
 
-  def self.require_for_daemon
-    require 'rpush/daemon'
-  end
-
   def self.logger
     @logger ||= Logger.new
+  end
+
+  def self.root
+    require 'rails'
+    Rails.root || Dir.pwd
+  rescue LoadError
+    Dir.pwd
   end
 
   class << self