+++ lib/rpush/daemon.rb
@@ -46,6 +46,8 @@ require 'rpush/daemon/adm'
 
 module Rpush
   module Daemon
+    extend Term::ANSIColor
+
     class << self
       attr_accessor :store
     end
@@ -60,6 +62,9 @@ module Rpush
       # No further store connections will be made from this thread.
       store.release_connection
 
+      Rpush.logger.info('Rpush operational.')
+      show_welcome_if_needed
+
       # Blocking call, returns after Feeder.stop is called from another thread.
       Feeder.start
 
@@ -68,12 +73,19 @@ module Rpush
     end
 
     def self.shutdown
-      puts "\nShutting down..."
+      if Rpush.config.foreground
+        # Eat the '^C'
+        STDOUT.write("\b\b")
+        STDOUT.flush
+      end
+
+      Rpush.logger.info('Shutting down... ', true)
 
       shutdown_lock.synchronize do
         Feeder.stop
         AppRunner.stop
         delete_pid_file
+        puts green('✔') if Rpush.config.foreground
       end
     end
 
@@ -105,6 +117,7 @@ module Rpush
     def self.write_pid_file
       unless Rpush.config.pid_file.blank?
         begin
+          FileUtils.mkdir_p(File.dirname(Rpush.config.pid_file))
           File.open(Rpush.config.pid_file, 'w') { |f| f.puts Process.pid }
         rescue SystemCallError => e
           Rpush.logger.error("Failed to write PID to '#{Rpush.config.pid_file}': #{e.inspect}")
@@ -115,6 +128,18 @@ module Rpush
     def self.delete_pid_file
       pid_file = Rpush.config.pid_file
       File.delete(pid_file) if !pid_file.blank? && File.exist?(pid_file)
+    end
+
+    def self.show_welcome_if_needed
+      if Rpush::Daemon::AppRunner.app_ids.count == 0
+        puts <<-EOS
+
+* #{green('Is this your first time using Rpush?')}
+  You need to create an App before you can start using Rpush.
+  Please refer to the documentation at https://github.com/rpush/rpush
+
+        EOS
+      end
     end
   end
 end