+++ lib/rpush/logger.rb
@@ -1,9 +1,8 @@
 module Rpush
   class Logger
     def initialize
-      log_dir = File.join(Rpush.config.log_dir, 'log')
-      FileUtils.mkdir_p(log_dir)
-      log = File.open(File.join(log_dir, 'rpush.log'), 'a')
+      FileUtils.mkdir_p(File.dirname(Rpush.config.log_file))
+      log = File.open(Rpush.config.log_file, 'a')
       log.sync = true
       setup_logger(log)
     rescue Errno::ENOENT, Errno::EPERM => e
@@ -12,16 +11,16 @@ module Rpush
       error('Logging disabled.')
     end
 
-    def info(msg)
-      log(:info, msg)
+    def info(msg, inline = false)
+      log(:info, msg, inline)
     end
 
-    def error(msg)
-      log(:error, msg, 'ERROR', STDERR)
+    def error(msg, inline = false)
+      log(:error, msg, inline, 'ERROR', STDERR)
     end
 
-    def warn(msg)
-      log(:warn, msg, 'WARNING', STDERR)
+    def warn(msg, inline = false)
+      log(:warn, msg, inline, 'WARNING', STDERR)
     end
 
     private
@@ -30,14 +29,22 @@ module Rpush
       if Rpush.config.logger
         @logger = Rpush.config.logger
       elsif ActiveSupport.const_defined?('BufferedLogger')
-        @logger = ActiveSupport::BufferedLogger.new(log, Rails.logger.level)
-        @logger.auto_flushing = Rails.logger.respond_to?(:auto_flushing) ? Rails.logger.auto_flushing : true
+        @logger = ActiveSupport::BufferedLogger.new(log, Rpush.config.log_level)
+        @logger.auto_flushing = auto_flushing
       else
-        @logger = ActiveSupport::Logger.new(log, Rails.logger.level)
+        @logger = ActiveSupport::Logger.new(log, Rpush.config.log_level)
       end
     end
 
-    def log(where, msg, prefix = nil, io = STDOUT)
+    def auto_flushing
+      if defined?(Rails) && Rails.logger.respond_to?(:auto_flushing)
+        Rails.logger.auto_flushing
+      else
+        true
+      end
+    end
+
+    def log(where, msg, inline = false, prefix = nil, io = STDOUT)
       if msg.is_a?(Exception)
         formatted_backtrace = msg.backtrace.join("\n")
         msg = "#{msg.class.name}, #{msg.message}\n#{formatted_backtrace}"
@@ -47,13 +54,19 @@ module Rpush
       formatted_msg << "[#{prefix}] " if prefix
       formatted_msg << msg
 
-      if io == STDERR
-        io.puts formatted_msg
-      elsif Rpush.config.foreground
-        io.puts formatted_msg
-      end
-
+      log_foreground(io, formatted_msg, inline)
       @logger.send(where, formatted_msg) if @logger
+    end
+
+    def log_foreground(io, formatted_msg, inline)
+      return unless io == STDERR || Rpush.config.foreground
+
+      if inline
+        io.write(formatted_msg)
+        io.flush
+      else
+        io.puts(formatted_msg)
+      end
     end
   end
 end