+++ CHANGELOG.md
@@ -2,6 +2,10 @@
   * Bump APNs max payload size to 2048 for iOS 8.
   * Add 'category' for iOS 8.
   * Add url_args for Safari Push Notification Support (#77).
+  * Improved command-line interface.
+  * Rails integration is now optional.
+  * Added log_level config option.
+  * log_dir is now deprecated and has no effect, use log_file instead.
 
 ## 2.0.1 (Sept 13, 2014)
   * Add ssl_certificate_revoked reflection (#68).